Better mouse
+ lots of new assets
+ broke out asset logic
+ Moduleized code
