+++ static/index.html
@@ -81,7 +81,7 @@
 					<span class="slider round"></span>
 				</label>
 				</div>
-				<div class="versionNum">RimMap v4.1 - @james_simo</div>
+				<div class="versionNum">RimMap v5.0 - @james_simo</div>
 			</div>
 
 
@@ -157,14 +157,14 @@
 				<div class="tileInfoContain">
 					<div id="tileData">
 						<ul>
-							<li><span id="terrainName"></span></li>
 							<li><span id="stuffName"></span></li>
+							<li><span id="terrainName"></span></li>
 							<li><span id="resourceName"></span></li>
 							<li><span id="deepResourceName"></span></li>
 						</ul>
 					</div>
 				</div>
-				<div class="versionNum">RimMap v4.1 - @james_simo</div>
+				<div class="versionNum">RimMap v5.0 - @james_simo</div>
 
 			</div>
 		</div>
